Refactoring for fxml structure
Added buttons to fxml and removed them from controller class
