+++ src/main/resources/fxml/sketch_on_map.fxml
@@ -23,10 +23,10 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<?import com.esri.arcgisruntime.mapping.view.MapView?>
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.paint.Color?>
-<StackPane fx:controller="com.esri.samples.displayinformation.sketch_on_map.SketchOnMapController" xmlns:fx="http://javafx.com/fxml"
+<StackPane fx:controller="com.esri.samples.displayinformation.sketch_on_map.SketchOnMapController"
+           xmlns:fx="http://javafx.com/fxml"
            stylesheets="/css/style.css">
     <MapView fx:id="mapView"/>
     <VBox StackPane.alignment="TOP_RIGHT" maxWidth="255" maxHeight="110" styleClass="panel-region" spacing="6"
@@ -45,11 +45,50 @@
                 </fills>
             </Background>
         </background>
+        <HBox maxWidth="200" maxHeight="50" spacing="6">
+            <Button fx:id="createPointButton" style="-fx-graphic: url('/icons/pointSketch.png')">
+                <tooltip>
+                    <Tooltip text="Sketch single point"></Tooltip>
+                </tooltip>
+            </Button>
+            <Button fx:id="createMultiPointButton" style="-fx-graphic: url('/icons/multiPointSketch.png')">
+                <tooltip>
+                    <Tooltip text="Sketch multipoints"></Tooltip>
+                </tooltip>
+            </Button>
+            <Button fx:id="createPolylineButton" style="-fx-graphic: url('/icons/polylineSketch.png')">
+                <tooltip>
+                    <Tooltip text="Sketch polyline"></Tooltip>
+                </tooltip>
+            </Button>
+            <Button fx:id="createPolygonButton" style="-fx-graphic: url('/icons/polygonSketch.png')">
+                <tooltip>
+                    <Tooltip text="Sketch Polygon"></Tooltip>
+                </tooltip>
+            </Button>
+            <Button fx:id="createFreehandPolylineButton" style="-fx-graphic: url('/icons/freehandPolylineSketch.png')">
+                <tooltip>
+                    <Tooltip text="Sketch freehand polyline"></Tooltip>
+                </tooltip>
+            </Button>
+            <Button fx:id="createFreehandPolygonButton" style="-fx-graphic: url('/icons/freehandPolygonSketch.png')">
+                <tooltip>
+                    <Tooltip text="Sketch freehand polygon"></Tooltip>
+                </tooltip>
+            </Button>
+        </HBox>
+        <Button fx:id="stopSketchButton" text="Sketch is disabled" maxWidth="Infinity">
+        </Button>
+        <HBox spacing="6" alignment="CENTER">
+            <Button fx:id="undoButton" text="Undo sketch"/>
+            <Button fx:id="redoButton" text="Redo sketch"/>
+        </HBox>
+        <Button fx:id="saveButton" text="Save sketch and stop sketching" maxWidth="Infinity"/>
+        <Button fx:id="editButton" text="Edit sketch" maxWidth="Infinity"/>
+        <Button fx:id="clearButton" text="Clear saved sketches" maxWidth="Infinity"/>
 
 
     </VBox>
-
-
 
 
 </StackPane>
